CSS and Images
I downloaded a css bootstrap for a fashion blog, I added in the fashion images to the images folder. I also copied the css into new css files in my project called mainStyles.css and variablesStyles.css. The css is not displaying with my project, that is what I will work on next.